+++ src/Products/Five/browser/pagetemplatefile.py
@@ -19,17 +19,25 @@ from os.path import basename
 from AccessControl import getSecurityManager
 from Acquisition import aq_get
 from Products.PageTemplates.Expressions import SecureModuleImporter
-from Products.PageTemplates.Expressions import createTrustedZopeEngine
+from Products.PageTemplates.Expressions import \
+     createTrustedZopeEngine as createPtTrustedZopeEngine
+from Products.PageTemplates.PageTemplate import get_template_engine_type
+from Products.PageTemplates.expression import \
+     createTrustedZopeEngine as createChTrustedZopeEngine
 from zope.component import getMultiAdapter
 from zope.pagetemplate.engine import TrustedAppPT
 from zope.pagetemplate.pagetemplatefile import PageTemplateFile
 
 
-_engine = createTrustedZopeEngine()
+_pt_engine = createPtTrustedZopeEngine()
+_ch_engine = createChTrustedZopeEngine()
 
 
 def getEngine():
-    return _engine
+    # we must delay the engine selection because the switch
+    #  from the `zope.pagetemplate` to the `chameleon` template engine
+    #  happens quite late (maybe after the import of this module)
+    return _pt_engine if get_template_engine_type() == "pt" else _ch_engine
 
 
 class ViewPageTemplateFile(TrustedAppPT, PageTemplateFile):